prepare: split out preparation code into separate files
This should be largely a non-functional change, though care should be
taken for the wasmparser version and such…

Hopefully it doesn’t change anything?
