+++ Cargo.toml
@@ -196,8 +196,7 @@ tracing-opentelemetry = "0.17.0"
 tracing-span-tree = "0.1"
 tracing-subscriber = { version = "0.3.15", features = ["env-filter", "fmt", "registry", "std"] }
 validator = "0.12"
-wasm-encoder = "0.11.0"
-wasm-smith = "0.10"
+wasm-smith = "0.12"
 wasmparser = "0.78"
 wasmprinter = "0.2"
 wasmtime = { version = "4.0.0", default-features = false, features = ["cranelift"] }
