+++ runtime/near-test-contracts/Cargo.toml
@@ -9,8 +9,8 @@ license = "Apache-2.0"
 [dependencies]
 once_cell = "1.5.0"
 wat = "1.0.40"
-wasm-encoder = "0.11.0"
-wasm-smith = "0.10.0"
+wasm-encoder = "0.22.0"
+wasm-smith = "0.12.0"
 rand = { version = "0.8.5", features = ["small_rng"] }
 arbitrary = "1.1.0"
